requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
datasets==2.5.1
dill==0.3.5.1
docker-pycreds==0.4.0
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.9.1
hydra-core==1.2.0
idna==3.4
more-itertools==8.14.0
multidict==6.0.2
multiprocess==0.70.13
numpy==1.23.3
omegaconf==2.2.3
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.20.2
psutil==5.9.2
pyarrow==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
responses==0.18.0
scipy==1.9.1
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tokenizers==0.12.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
urllib3==1.26.12
wandb==0.13.3
xxhash==3.0.0
yarl==1.8.1
